--- src/components/forms/dialog-forms/dialog-image-delete/index.ts
@@ -1,1 +1,0 @@
-export * from './dialog-image-delete'
